pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "trafilatura>=2.0.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.3",
    "flask-migrate>=4.1.0",
    "pillow>=11.1.0",
]
